add missing rule for stuttgart table
Previously, we had rules for the stuttgart table only for packets
arriving on brXX. But that's not enough because this packets might
also reach us over backbone or internet.

Hence, just get rid of this and replace it by a single rule:
  7000:	from all lookup stuttgart

Because there currently is no mechanism for globally adding things, we
do this in the segment interfaces (TODO build something like backbones
/etc/systemd/system/ffs-ip6-policy.service). We don't delete this
rule so it's not removed when we bring down a single client bridge.